requirements.txt
aiogram==3.17.0
aiohttp==3.11.10
aiosqlite==0.20.0
# pydub==0.25.1 # Olib tashlandi (kerak emas)
python-dotenv==1.0.1
SQLAlchemy==2.0.36
youtube-search-python==1.6.6

lyricsgenius==3.0.1
beautifulsoup4==4.12.3
fuzzywuzzy==0.18.0
# python-Levenshtein==0.23.0
